+++ libs/graphql-schema/.graphqlrc.ts
@@ -1,0 +1,4 @@
+export default {
+  schema: 'generated/schema.graphql',
+  documents: 'src/apollo/*.gql',
+}